+++ tidy3d/plugins/autograd/__init__.py
@@ -1,0 +1,53 @@
+from .functions import (
+    convolve,
+    grey_closing,
+    grey_dilation,
+    grey_erosion,
+    grey_opening,
+    morphological_gradient,
+    morphological_gradient_external,
+    morphological_gradient_internal,
+    pad,
+    rescale,
+    threshold,
+)
+from .invdes import (
+    grey_indicator,
+    make_circular_filter,
+    make_conic_filter,
+    make_curvature_penalty,
+    make_erosion_dilation_penalty,
+    make_filter,
+    make_filter_and_project,
+    ramp_projection,
+    tanh_projection,
+)
+from .primitives import gaussian_filter
+from .utilities import chain, get_kernel_size_px, make_kernel
+
+__all__ = [
+    "make_filter",
+    "make_conic_filter",
+    "make_circular_filter",
+    "grey_indicator",
+    "convolve",
+    "pad",
+    "ramp_projection",
+    "tanh_projection",
+    "make_erosion_dilation_penalty",
+    "make_curvature_penalty",
+    "make_filter_and_project",
+    "gaussian_filter",
+    "make_kernel",
+    "get_kernel_size_px",
+    "chain",
+    "grey_closing",
+    "grey_dilation",
+    "grey_erosion",
+    "grey_opening",
+    "morphological_gradient",
+    "morphological_gradient_external",
+    "morphological_gradient_internal",
+    "rescale",
+    "threshold",
+]