+++ tidy3d/plugins/autograd/invdes/parametrizations.py
@@ -1,4 +1,4 @@
-from typing import Tuple
+from typing import Tuple, Union
 
 import autograd.numpy as np
 
@@ -8,7 +8,10 @@ from .projections import tanh_projection
 
 
 def make_filter_and_project(
-    filter_size: Tuple[int, ...],
+    radius: Union[float, Tuple[float, ...]] = None,
+    dl: Union[float, Tuple[float, ...]] = None,
+    *,
+    size_px: Union[int, Tuple[int, ...]] = None,
     beta: float = 1.0,
     eta: float = 0.5,
     filter_type: KernelType = "conic",
@@ -20,8 +23,12 @@ def make_filter_and_project(
 
     Parameters
     ----------
-    filter_size : Tuple[int, ...]
-        The size of the filter kernel in pixels.
+    radius : Union[float, Tuple[float, ...]], optional
+        The radius of the kernel. Can be a scalar or a tuple. Default is None.
+    dl : Union[float, Tuple[float, ...]], optional
+        The grid spacing. Can be a scalar or a tuple. Default is None.
+    size_px : Union[int, Tuple[int, ...]], optional
+        The size of the kernel in pixels for each dimension. Can be a scalar or a tuple. Default is None.
     beta : float, optional
         The beta parameter for the tanh projection, by default 1.0.
     eta : float, optional
@@ -36,7 +43,7 @@ def make_filter_and_project(
     function
         A function that takes an array and applies the filter and projection.
     """
-    _filter = make_filter(filter_type, filter_size, padding=padding)
+    _filter = make_filter(radius, dl, size_px=size_px, filter_type=filter_type, padding=padding)
 
     def _filter_and_project(array: np.ndarray, beta: float = beta, eta: float = eta):
         array = _filter(array)